simple_ai_agent_with_gradio/requirements.txt
langchain==1.2.10
google-genai==1.65.0
langchain-google-genai==4.2.1
ollama==0.6.1
langchain-ollama==1.0.1
langchain-community==0.4.1
gradio==6.9.0
langgraph==1.0.10
langgraph-checkpoint-sqlite==3.0.3
python-dotenv==1.2.2

